🚀 [Feat]: LabelType => Label 변경
